requirements.txt
django
mysqlclient # mysql support
django-mysql
dateparser

django-guardian
django-bootstrap4
django-crispy-forms
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support

Json-Dict